Working on Windows/MSC build

+++ src/framework/ImageFile.cpp
@@ -6,7 +6,7 @@
 ImageFile::ImageFile(std::string_view filename, int components) {
   auto res = get_resource(filename);
   m_data = {
-    stbi_load_from_memory((const stbi_uc*)res.first, res.second,
+    stbi_load_from_memory((const stbi_uc*)res.first, static_cast<int>(res.second),
       &m_width, &m_height, &m_components, components),
     [](void* handle) { stbi_image_free(handle); }
   };
